Change get comment sentiment from load comments servlet to new comment servlet
