+++ portfolio/src/main/java/com/google/sps/servlets/LoadCommentServlet.java
@@ -20,9 +20,6 @@ import com.google.appengine.api.datastore.DatastoreService;
 import com.google.appengine.api.datastore.DatastoreServiceFactory;
 import com.google.appengine.api.datastore.Entity;
 import com.google.appengine.api.datastore.PreparedQuery;
-import com.google.cloud.language.v1.Document;
-import com.google.cloud.language.v1.LanguageServiceClient;
-import com.google.cloud.language.v1.Sentiment;
 import com.google.cloud.translate.Translate;
 import com.google.cloud.translate.TranslateOptions;
 import com.google.cloud.translate.Translation;
@@ -60,19 +57,6 @@ public class LoadCommentServlet extends HttpServlet {
     return translatedText;
   }
 
-  // Get the comment sentiment using Google's Sentiment Analysis API
-  public float getCommentSentiment(String comment) throws IOException{
-    Document sentimentDoc =
-      Document.newBuilder().setContent(comment).setType(Document.Type.PLAIN_TEXT).build();
-    LanguageServiceClient languageService = LanguageServiceClient.create();
-    Sentiment sentiment = languageService.analyzeSentiment(sentimentDoc).getDocumentSentiment();
-    float sentimentScore = sentiment.getScore();
-
-    languageService.close();
-
-    return sentimentScore;
-  }
-
   public void loadComments() throws IOException {
     Query commentsQuery = new Query("Comment"); // Get previous stored comments
     commentArray.clear(); // Empty the array on every comments GET.
@@ -89,8 +73,8 @@ public class LoadCommentServlet extends HttpServlet {
       String comment = (String) commentEntity.getProperty("comment");
 
       String translatedComment = translateComment(comment);
-      float sentimentScore = getCommentSentiment(comment);
 
+      double sentimentScore = (double)commentEntity.getProperty("sentiment-score");
       String user = (String) commentEntity.getProperty("user");
       String email = (String) commentEntity.getProperty("email");
       String userId = (String) commentEntity.getProperty("userId");
